retina tweaks, renames, euify
