+++ src-docs/src/views/icon/icons.js
@@ -58,9 +58,9 @@ const iconTypes = [
   'cut',
   'database',
   'document',
+  'documentEdit',
   'documents',
   'dot',
-  'edit',
   'email',
   'empty',
   'exit',
