--- src/components/icon/assets/cloud-drizzle.js
@@ -1,17 +1,0 @@
-import React from 'react';
-
-const EuiIconCloudDrizzle = props => (
-  <svg
-    width={16}
-    height={16}
-    viewBox="0 0 16 16"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}>
-    <g fill="#000" fillRule="nonzero">
-      <path d="M6.348 3.761A3.995 3.995 0 0 1 8 7a.5.5 0 0 1-1 0 3 3 0 1 0-4.878 2.34.5.5 0 0 1-.627.779 4 4 0 0 1 3.973-6.84 5.502 5.502 0 0 1 10.096 4.37.5.5 0 1 1-.92-.39 4.5 4.5 0 0 0-8.296-3.497z" />
-      <path d="M4.739 8.696l-1.466 1.792a.5.5 0 0 0 .774.633L5.513 9.33a.5.5 0 1 0-.774-.633zm-3.12 3.647l-1.506 1.84a.5.5 0 1 0 .774.634l1.505-1.84a.5.5 0 1 0-.774-.634zm7.62-3.647l-1.466 1.792a.5.5 0 0 0 .774.633l1.466-1.792a.5.5 0 0 0-.774-.633zm-3.12 3.647l-1.506 1.84a.5.5 0 1 0 .774.634l1.505-1.84a.5.5 0 1 0-.774-.634zm7.62-3.647l-1.466 1.792a.5.5 0 1 0 .774.633l1.466-1.792a.5.5 0 0 0-.774-.633zm-3.12 3.647l-1.506 1.84a.5.5 0 1 0 .774.634l1.505-1.84a.5.5 0 1 0-.774-.634z" />
-    </g>
-  </svg>
-);
-
-export const icon = EuiIconCloudDrizzle;
