--- src/components/icon/assets/cloud-stormy.js
@@ -1,17 +1,0 @@
-import React from 'react';
-
-const EuiIconCloudStormy = props => (
-  <svg
-    width={16}
-    height={16}
-    viewBox="0 0 16 16"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}>
-    <g fill="#000" fillRule="nonzero">
-      <path d="M5 12a.5.5 0 0 1-.343-.864l4.25-4c.454-.428 1.124.195.731.68L7.05 11H9.5a.5.5 0 0 1 .32.885l-4.822 4c-.484.402-1.116-.275-.681-.73L7.331 12H5z" />
-      <path d="M6.348 3.761c.562.408 1.018.96 1.31 1.617a.5.5 0 1 1-.915.406 3 3 0 1 0-3.579 4.098.5.5 0 1 1-.277.96 4.001 4.001 0 1 1 2.58-7.564 5.5 5.5 0 1 1 6.127 7.613.5.5 0 1 1-.198-.98 4.502 4.502 0 1 0-5.048-6.15zm-4.55 8.362l-1.685 2.06a.5.5 0 1 0 .774.634l1.686-2.06a.5.5 0 0 0-.774-.634zm10 0l-1.685 2.06a.5.5 0 1 0 .774.634l1.686-2.06a.5.5 0 1 0-.774-.634z" />
-    </g>
-  </svg>
-);
-
-export const icon = EuiIconCloudStormy;
