+++ src/components/icon/assets/cloudStormy.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const EuiIconCloudStormy = props => (
+  <svg
+    width={16}
+    height={16}
+    viewBox="0 0 16 16"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}>
+    <path d="M9.6 11c.669.002.794.67.36 1.003l-4.68 3.882c-.457.378-1.053-.26-.643-.689l3.08-3.193A5411.7 5411.7 0 0 1 5.113 12c-.668-.001-.793-.669-.36-1.003l4.68-3.881c.458-.379 1.053.26.643.688l-3.08 3.193L9.6 11zM5.5 3.28a5.5 5.5 0 1 1 7.003 7.357.5.5 0 1 1-.36-.934 4.5 4.5 0 1 0-5.797-5.856c.36.338.654.749.86 1.214a.47.47 0 1 1-.86.382 2.824 2.824 0 1 0-3.368 3.858.47.47 0 1 1-.261.904 3.766 3.766 0 1 1 2.429-7.12l.354.195z" />
+  </svg>
+);
+
+export const icon = EuiIconCloudStormy;
