+++ src/components/icon/assets/documents.js
@@ -7,10 +7,7 @@ const EuiIconDocuments = props => (
     viewBox="0 0 16 16"
     xmlns="http://www.w3.org/2000/svg"
     {...props}>
-    <g fill="#000" fillRule="nonzero">
-      <path d="M8 1H2v11h10V5H8.5a.5.5 0 0 1-.5-.5V1zm1 0v3h3L9 1zM1 1a1 1 0 0 1 1-1h7a1 1 0 0 1 .707.293l3 3A1 1 0 0 1 13 4v8a1 1 0 0 1-1 1H2a1 1 0 0 1-1-1V1z" />
-      <path d="M4.5 8h5a.5.5 0 0 0 0-1h-5a.5.5 0 0 0 0 1zm0 2h5a.5.5 0 0 0 0-1h-5a.5.5 0 0 0 0 1zM6 16h9a1 1 0 0 0 1-1V6a.5.5 0 1 0-1 0v9H6a.5.5 0 1 0 0 1z" />
-    </g>
+    <path d="M8.8 0c.274 0 .537.113.726.312l3.2 3.428c.176.186.274.433.274.689V13a1 1 0 0 1-1 1H2a1 1 0 0 1-1-1V1a1 1 0 0 1 1-1h6.8zM12 5H8.5a.5.5 0 0 1-.5-.5V1H2v12h10V5zm-7.5 6a.5.5 0 1 1 0-1h5a.5.5 0 1 1 0 1h-5zm0-3a.5.5 0 0 1 0-1h5a.5.5 0 1 1 0 1h-5zm1 8a.5.5 0 1 1 0-1H14V6.5a.5.5 0 1 1 1 0V15a1 1 0 0 1-1 1H5.5z" />
   </svg>
 );
