--- src/components/icon/assets/edit.js
@@ -1,26 +1,0 @@
-import React from 'react';
-
-const EuiIconEdit = props => (
-  <svg
-    width={16}
-    height={16}
-    viewBox="0 0 16 16"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}>
-    <g
-      id="icon-blog-16-ink"
-      stroke="none"
-      strokeWidth={1}
-      fill="#343741"
-      fillRule="nonzero">
-      <path
-        d="M7 6.293V9h2.707l5.44-5.44a1.5 1.5 0 0 0 0-2.12L14.56.853a1.5 1.5 0 0 0-2.122 0L7 6.293zm1 .414l5.146-5.146a.5.5 0 0 1 .708 0l.585.585a.5.5 0 0 1 0 .708L9.293 8H8V6.707z"
-        id="Path-7"
-      />
-      <path id="Path-9" d="M4 13h8v-1H4z" />
-      <path id="Path-8" d="M14 15H2V2h7V1H1v15h14V7h-1z" />
-    </g>
-  </svg>
-);
-
-export const icon = EuiIconEdit;
