+++ src/components/icon/icon.tsx
@@ -45,9 +45,9 @@ const typeToPathMap = {
   check: 'check',
   checkInCircleFilled: 'checkInCircleFilled',
   clock: 'clock',
-  cloudDrizzle: 'cloud-drizzle',
-  cloudStormy: 'cloud-stormy',
-  cloudSunny: 'cloud-sunny',
+  cloudDrizzle: 'cloudDrizzle',
+  cloudStormy: 'cloudStormy',
+  cloudSunny: 'cloudSunny',
   compute: 'compute',
   console: 'console',
   consoleApp: 'app_console',
@@ -70,9 +70,9 @@ const typeToPathMap = {
   devToolsApp: 'app_devtools',
   discoverApp: 'app_discover',
   document: 'document',
+  documentEdit: 'documentEdit',
   documents: 'documents',
   dot: 'dot',
-  edit: 'edit',
   editorAlignCenter: 'editor_align_center',
   editorAlignLeft: 'editor_align_left',
   editorAlignRight: 'editor_align_right',
@@ -292,7 +292,7 @@ const typeToPathMap = {
   user: 'user',
   usersRolesApp: 'app_users_roles',
   vector: 'vector',
-  videoPlayer: 'video-player',
+  videoPlayer: 'videoPlayer',
   visArea: 'vis_area',
   visBarHorizontal: 'vis_bar_horizontal',
   visBarVertical: 'vis_bar_vertical',
